refactor: consolidate shopping layout and remove unused components
- Merged ShoppingLayout into the main layout, integrating ShoppingCartProvider and CartSidebar directly.
- Updated ShoppingPage to render Marketplace component.
- Deleted redundant shopping layout and payment page files to streamline the codebase.
