+++ src/app/payment/page.tsx
@@ -2,7 +2,17 @@
 
 import React, { useState } from "react";
 import { useRouter } from "next/navigation";
-import { ArrowLeft, CreditCard, Shield, CheckCircle, XCircle, User, MapPin, ChevronRight, ChevronLeft } from "lucide-react";
+import {
+  ArrowLeft,
+  CreditCard,
+  Shield,
+  CheckCircle,
+  XCircle,
+  User,
+  MapPin,
+  ChevronRight,
+  ChevronLeft,
+} from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
@@ -40,26 +50,26 @@ const paymentSteps: PaymentStep[] = [
     id: 1,
     title: "Contact Information",
     description: "Enter your email address",
-    icon: User
+    icon: User,
   },
   {
     id: 2,
     title: "Shipping Details",
     description: "Your delivery address",
-    icon: MapPin
+    icon: MapPin,
   },
   {
     id: 3,
     title: "Payment Method",
     description: "Card details and confirmation",
-    icon: CreditCard
-  }
+    icon: CreditCard,
+  },
 ];
 
 export default function PaymentPage() {
   const router = useRouter();
   const { cartItems, cartTotal, clearCart } = useShoppingCart();
-  
+
   const [currentStep, setCurrentStep] = useState(1);
   const [formData, setFormData] = useState<PaymentFormData>({
     email: "",
@@ -75,7 +85,9 @@ export default function PaymentPage() {
   });
 
   const [isProcessing, setIsProcessing] = useState(false);
-  const [paymentResult, setPaymentResult] = useState<PaymentResult | null>(null);
+  const [paymentResult, setPaymentResult] = useState<PaymentResult | null>(
+    null
+  );
   const [errors, setErrors] = useState<Partial<PaymentFormData>>({});
   const [stepErrors, setStepErrors] = useState<{ [key: number]: boolean }>({});
 
@@ -83,10 +95,10 @@ export default function PaymentPage() {
   const finalTotal = cartTotal + taxAmount;
 
   const handleInputChange = (field: keyof PaymentFormData, value: string) => {
-    setFormData(prev => ({ ...prev, [field]: value }));
+    setFormData((prev) => ({ ...prev, [field]: value }));
     // Clear error when user starts typing
     if (errors[field]) {
-      setErrors(prev => ({ ...prev, [field]: undefined }));
+      setErrors((prev) => ({ ...prev, [field]: undefined }));
     }
   };
 
@@ -158,7 +170,7 @@ export default function PaymentPage() {
     }
 
     setErrors(newErrors);
-    setStepErrors(prev => ({ ...prev, [step]: !isValid }));
+    setStepErrors((prev) => ({ ...prev, [step]: !isValid }));
     return isValid;
   };
 
@@ -168,17 +180,17 @@ export default function PaymentPage() {
 
   const nextStep = () => {
     if (validateStep(currentStep)) {
-      setCurrentStep(prev => Math.min(prev + 1, 3));
+      setCurrentStep((prev) => Math.min(prev + 1, 3));
     }
   };
 
   const prevStep = () => {
-    setCurrentStep(prev => Math.max(prev - 1, 1));
+    setCurrentStep((prev) => Math.max(prev - 1, 1));
   };
 
   const simulatePaymentGateway = async (): Promise<PaymentResult> => {
     // Simulate network delay
-    await new Promise(resolve => setTimeout(resolve, 2000));
+    await new Promise((resolve) => setTimeout(resolve, 2000));
 
     // Simulate random success/failure for demo (90% success rate)
     const isSuccess = Math.random() > 0.1;
@@ -187,19 +199,20 @@ export default function PaymentPage() {
       return {
         success: true,
         transactionId: `TXN${Date.now()}${Math.floor(Math.random() * 1000)}`,
-        message: "Payment processed successfully!"
+        message: "Payment processed successfully!",
       };
     } else {
       return {
         success: false,
-        message: "Payment failed. Please check your card details and try again."
+        message:
+          "Payment failed. Please check your card details and try again.",
       };
     }
   };
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
-    
+
     if (currentStep < 3) {
       nextStep();
       return;
@@ -210,22 +223,23 @@ export default function PaymentPage() {
     }
 
     setIsProcessing(true);
-    
+
     try {
       const result = await simulatePaymentGateway();
       setPaymentResult(result);
-      
+
       if (result.success) {
         // Clear cart on successful payment
         setTimeout(() => {
           clearCart();
-          router.push("/shopping");
+          router.push("/");
         }, 3000);
       }
     } catch {
       setPaymentResult({
         success: false,
-        message: "An error occurred while processing your payment. Please try again."
+        message:
+          "An error occurred while processing your payment. Please try again.",
       });
     } finally {
       setIsProcessing(false);
@@ -266,11 +280,15 @@ export default function PaymentPage() {
                 <Input
                   id="firstName"
                   value={formData.firstName}
-                  onChange={(e) => handleInputChange("firstName", e.target.value)}
+                  onChange={(e) =>
+                    handleInputChange("firstName", e.target.value)
+                  }
                   className={errors.firstName ? "border-destructive" : ""}
                 />
                 {errors.firstName && (
-                  <p className="text-sm text-destructive mt-1">{errors.firstName}</p>
+                  <p className="text-sm text-destructive mt-1">
+                    {errors.firstName}
+                  </p>
                 )}
               </div>
               <div>
@@ -278,11 +296,15 @@ export default function PaymentPage() {
                 <Input
                   id="lastName"
                   value={formData.lastName}
-                  onChange={(e) => handleInputChange("lastName", e.target.value)}
+                  onChange={(e) =>
+                    handleInputChange("lastName", e.target.value)
+                  }
                   className={errors.lastName ? "border-destructive" : ""}
                 />
                 {errors.lastName && (
-                  <p className="text-sm text-destructive mt-1">{errors.lastName}</p>
+                  <p className="text-sm text-destructive mt-1">
+                    {errors.lastName}
+                  </p>
                 )}
               </div>
             </div>
@@ -296,7 +318,9 @@ export default function PaymentPage() {
                 className={errors.address ? "border-destructive" : ""}
               />
               {errors.address && (
-                <p className="text-sm text-destructive mt-1">{errors.address}</p>
+                <p className="text-sm text-destructive mt-1">
+                  {errors.address}
+                </p>
               )}
             </div>
             <div className="grid grid-cols-1 md:grid-cols-2 gap-4">
@@ -319,11 +343,15 @@ export default function PaymentPage() {
                   id="postalCode"
                   placeholder="00100"
                   value={formData.postalCode}
-                  onChange={(e) => handleInputChange("postalCode", e.target.value)}
+                  onChange={(e) =>
+                    handleInputChange("postalCode", e.target.value)
+                  }
                   className={errors.postalCode ? "border-destructive" : ""}
                 />
                 {errors.postalCode && (
-                  <p className="text-sm text-destructive mt-1">{errors.postalCode}</p>
+                  <p className="text-sm text-destructive mt-1">
+                    {errors.postalCode}
+                  </p>
                 )}
               </div>
             </div>
@@ -342,11 +370,15 @@ export default function PaymentPage() {
                 id="cardholderName"
                 placeholder="John Doe"
                 value={formData.cardholderName}
-                onChange={(e) => handleInputChange("cardholderName", e.target.value)}
+                onChange={(e) =>
+                  handleInputChange("cardholderName", e.target.value)
+                }
                 className={errors.cardholderName ? "border-destructive" : ""}
               />
               {errors.cardholderName && (
-                <p className="text-sm text-destructive mt-1">{errors.cardholderName}</p>
+                <p className="text-sm text-destructive mt-1">
+                  {errors.cardholderName}
+                </p>
               )}
             </div>
             <div>
@@ -355,12 +387,19 @@ export default function PaymentPage() {
                 id="cardNumber"
                 placeholder="1234 5678 9012 3456"
                 value={formData.cardNumber}
-                onChange={(e) => handleInputChange("cardNumber", formatCardNumber(e.target.value))}
+                onChange={(e) =>
+                  handleInputChange(
+                    "cardNumber",
+                    formatCardNumber(e.target.value)
+                  )
+                }
                 maxLength={19}
                 className={errors.cardNumber ? "border-destructive" : ""}
               />
               {errors.cardNumber && (
-                <p className="text-sm text-destructive mt-1">{errors.cardNumber}</p>
+                <p className="text-sm text-destructive mt-1">
+                  {errors.cardNumber}
+                </p>
               )}
             </div>
             <div className="grid grid-cols-2 gap-4">
@@ -370,12 +409,19 @@ export default function PaymentPage() {
                   id="expiryDate"
                   placeholder="MM/YY"
                   value={formData.expiryDate}
-                  onChange={(e) => handleInputChange("expiryDate", formatExpiryDate(e.target.value))}
+                  onChange={(e) =>
+                    handleInputChange(
+                      "expiryDate",
+                      formatExpiryDate(e.target.value)
+                    )
+                  }
                   maxLength={5}
                   className={errors.expiryDate ? "border-destructive" : ""}
                 />
                 {errors.expiryDate && (
-                  <p className="text-sm text-destructive mt-1">{errors.expiryDate}</p>
+                  <p className="text-sm text-destructive mt-1">
+                    {errors.expiryDate}
+                  </p>
                 )}
               </div>
               <div>
@@ -384,7 +430,9 @@ export default function PaymentPage() {
                   id="cvv"
                   placeholder="123"
                   value={formData.cvv}
-                  onChange={(e) => handleInputChange("cvv", e.target.value.replace(/\D/g, ""))}
+                  onChange={(e) =>
+                    handleInputChange("cvv", e.target.value.replace(/\D/g, ""))
+                  }
                   maxLength={4}
                   className={errors.cvv ? "border-destructive" : ""}
                 />
@@ -410,7 +458,7 @@ export default function PaymentPage() {
   const formatCardNumber = (value: string) => {
     const v = value.replace(/\s+/g, "").replace(/[^0-9]/gi, "");
     const matches = v.match(/\d{4,16}/g);
-    const match = matches && matches[0] || "";
+    const match = (matches && matches[0]) || "";
     const parts = [];
     for (let i = 0, len = match.length; i < len; i += 4) {
       parts.push(match.substring(i, i + 4));
@@ -440,11 +488,10 @@ export default function PaymentPage() {
               No Items to Checkout
             </h2>
             <p className="text-muted-foreground mb-4">
-              Your cart is empty. Add some products before proceeding to payment.
+              Your cart is empty. Add some products before proceeding to
+              payment.
             </p>
-            <Button onClick={() => router.push("/shopping")}>
-              Back to Shopping
-            </Button>
+            <Button onClick={() => router.push("/")}>Back to Shopping</Button>
           </CardContent>
         </Card>
       </div>
@@ -546,7 +593,11 @@ export default function PaymentPage() {
                           <div className="mt-2 text-center">
                             <p
                               className={`text-sm font-medium ${
-                                isActive ? "text-primary" : hasError ? "text-destructive" : "text-muted-foreground"
+                                isActive
+                                  ? "text-primary"
+                                  : hasError
+                                  ? "text-destructive"
+                                  : "text-muted-foreground"
                               }`}
                             >
                               {step.title}
@@ -632,7 +683,10 @@ export default function PaymentPage() {
                 {/* Cart Items */}
                 <div className="space-y-3">
                   {cartItems.map((item) => (
-                    <div key={item.id} className="flex justify-between items-start">
+                    <div
+                      key={item.id}
+                      className="flex justify-between items-start"
+                    >
                       <div className="flex-1">
                         <h4 className="text-sm font-medium text-foreground">
                           {item.name}
@@ -651,7 +705,9 @@ export default function PaymentPage() {
                 <div className="border-t border-border pt-4 space-y-2">
                   <div className="flex justify-between text-sm">
                     <span className="text-muted-foreground">Subtotal:</span>
-                    <span className="text-foreground">LKR {(cartTotal * 365).toFixed(2)}</span>
+                    <span className="text-foreground">
+                      LKR {(cartTotal * 365).toFixed(2)}
+                    </span>
                   </div>
                   <div className="flex justify-between text-sm">
                     <span className="text-muted-foreground">Shipping:</span>
@@ -659,12 +715,16 @@ export default function PaymentPage() {
                   </div>
                   <div className="flex justify-between text-sm">
                     <span className="text-muted-foreground">Tax (10%):</span>
-                    <span className="text-foreground">LKR {(taxAmount * 365).toFixed(2)}</span>
+                    <span className="text-foreground">
+                      LKR {(taxAmount * 365).toFixed(2)}
+                    </span>
                   </div>
                   <div className="border-t border-border pt-2">
                     <div className="flex justify-between font-semibold text-lg">
                       <span className="text-foreground">Total:</span>
-                      <span className="text-primary">LKR {(finalTotal * 365).toFixed(2)}</span>
+                      <span className="text-primary">
+                        LKR {(finalTotal * 365).toFixed(2)}
+                      </span>
                     </div>
                   </div>
                 </div>
@@ -674,10 +734,12 @@ export default function PaymentPage() {
                   <div className="flex items-start space-x-2">
                     <Shield className="h-4 w-4 text-primary mt-0.5" />
                     <div>
-                      <p className="text-xs font-medium text-foreground">Secure Payment</p>
+                      <p className="text-xs font-medium text-foreground">
+                        Secure Payment
+                      </p>
                       <p className="text-xs text-muted-foreground">
-                        Your payment information is encrypted and processed securely. 
-                        We accept all major credit and debit cards.
+                        Your payment information is encrypted and processed
+                        securely. We accept all major credit and debit cards.
                       </p>
                     </div>
                   </div>
@@ -690,7 +752,7 @@ export default function PaymentPage() {
                     <span>{Math.round((currentStep / 3) * 100)}% Complete</span>
                   </div>
                   <div className="w-full bg-muted rounded-full h-2">
-                    <div 
+                    <div
                       className="bg-primary h-2 rounded-full transition-all duration-300"
                       style={{ width: `${(currentStep / 3) * 100}%` }}
                     />
