--- src/app/shopping/page.tsx
@@ -1,5 +1,0 @@
-import Marketplace from "@/components/shopping/Marketplace";
-
-export default function ShoppingPage() {
-  return <Marketplace />;
-}
